Fix minor version check for PermGen
The check for the PermGen to apply the flags for Java versions less than 8 was incorrect. The correct operator is LSS, not LEQ (which is Less Than or Equal to) and the value should not be quoted in this case.

As it was, the check would always fail, thus doing the right thing for Java 8 but 6 and 7 would have JMeter run without this flag set. I ran into this problem with a client's server. Normally we'd use a Linux box but we don't always get what we want in this life.